+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- **New data source:**  `scalr_module_versions` ([360](https://github.com/Scalr/terraform-provider-scalr/pull/360))
 - `data.scalr_variable`: new attributes `updated_at`, `updated_by` and `updated_by_email` ([354](https://github.com/Scalr/terraform-provider-scalr/pull/354))
 - `scalr_variable`: new attributes `updated_at`, `updated_by` and `updated_by_email` ([354](https://github.com/Scalr/terraform-provider-scalr/pull/354))